requirements.txt
# --- Core Framework ---
langchain>=0.3.0
langchain-ollama>=0.1.0     
langchain-experimental>=0.3.0 
langchain-google-genai>=2.0.0

# --- Data Handling ---
pandas>=2.0.0
numpy>=1.24.0

# --- Environment ---
python-dotenv>=1.0.0         

# ---  Visualization  ---
matplotlib>=3.7.0
seaborn>=0.12.0